엔티티 페치방식 수정
- 모든 엔티티 **ToOne 에서 fetch=FetchType.LAZY 설정
- 핀 상태 자료형 long으로 변경

+++ src/main/java/net/softbell/bsh/dto/view/advance/NodeItemCardDto.java
@@ -23,7 +23,7 @@ public class NodeItemCardDto
 	private PinTypeRule pinType;
 	private PinModeRule pinMode;
 	private byte controlMode;
-	private short lastStatus;
+	private long lastStatus;
 	
 	public NodeItemCardDto(NodeItem entity)
 	{
